Listen please.
I added a only-vines filter for the region min="oo" max= "-oo", so the only blocks that can be placed in the entire world are vines. No need to include extra unnecessary filter or regions thus.

I included a bow, which I thought you wanted (with a 1-arrow killreward). The Bow is included on Line 39, the start Arrows are included in Line 40, and the killreward of 1 Arrow per kill is included in Line 82 - 84.